+++ src/engine/economy/index.ts
@@ -7,3 +7,4 @@ export * from './playerToken';
 export * from './bank';
 export * from './cashSwipe';
 export * from './unemploymentCheck';
+export * from './netWorth';